requirements.txt
streamlit
pandas
scikit-learn
rdkit-pypi
numpy
biopython
